requirements.txt
aiohttp>=3.9.0
pandas>=2.0.0
pandas-ta>=0.3.14b
python-telegram-bot>=20.0
python-dotenv>=1.0.0
